Add role picker system
Adds a role picker system managed using the /role-picker command.
A role picker is a button that can be clicked to bring up a dropdown to select new roles. This two step process is done so defaults can be set to match the user's current roles.

Each role picker is registered as the button `luna_role_picker_{id}` and a global listener takes care of them. I sadly don't think there is a better way to go about it, due to disnake API limitations.